+++ ray-operator/test/sampleyaml/support.go
@@ -2,7 +2,6 @@ package sampleyaml
 
 import (
 	"os"
-	"os/exec"
 	"path/filepath"
 	"runtime"
 
@@ -11,11 +10,10 @@ import (
 	corev1 "k8s.io/api/core/v1"
 
 	rayv1 "github.com/ray-project/kuberay/ray-operator/apis/ray/v1"
-	rayscheme "github.com/ray-project/kuberay/ray-operator/pkg/client/clientset/versioned/scheme"
 	. "github.com/ray-project/kuberay/ray-operator/test/support"
 )
 
-func getSampleYAMLDir(t Test) string {
+func GetSampleYAMLDir(t Test) string {
 	t.T().Helper()
 	_, b, _, _ := runtime.Caller(0)
 	sampleYAMLDir := filepath.Join(filepath.Dir(b), "../../config/samples")
@@ -23,55 +21,6 @@ func getSampleYAMLDir(t Test) string {
 	assert.NoError(t.T(), err)
 	assert.True(t.T(), info.IsDir())
 	return sampleYAMLDir
-}
-
-func readYAML(t Test, filename string) []byte {
-	t.T().Helper()
-	sampleYAMLDir := getSampleYAMLDir(t)
-	yamlFile := filepath.Join(sampleYAMLDir, filename)
-	yamlFileContent, err := os.ReadFile(yamlFile)
-	assert.NoError(t.T(), err)
-	return yamlFileContent
-}
-
-func DeserializeRayClusterSampleYAML(t Test, filename string) *rayv1.RayCluster {
-	t.T().Helper()
-	yamlFileContent := readYAML(t, filename)
-	decoder := rayscheme.Codecs.UniversalDecoder()
-	rayCluster := &rayv1.RayCluster{}
-	_, _, err := decoder.Decode(yamlFileContent, nil, rayCluster)
-	assert.NoError(t.T(), err)
-	return rayCluster
-}
-
-func DeserializeRayServiceSampleYAML(t Test, filename string) *rayv1.RayService {
-	t.T().Helper()
-	yamlFileContent := readYAML(t, filename)
-	decoder := rayscheme.Codecs.UniversalDecoder()
-	rayService := &rayv1.RayService{}
-	_, _, err := decoder.Decode(yamlFileContent, nil, rayService)
-	assert.NoError(t.T(), err)
-	return rayService
-}
-
-func DeserializeRayJobSampleYAML(t Test, filename string) *rayv1.RayJob {
-	t.T().Helper()
-	yamlFileContent := readYAML(t, filename)
-	decoder := rayscheme.Codecs.UniversalDecoder()
-	rayJob := &rayv1.RayJob{}
-	_, _, err := decoder.Decode(yamlFileContent, nil, rayJob)
-	assert.NoError(t.T(), err)
-	return rayJob
-}
-
-func KubectlApplyYAML(t Test, filename string, namespace string) {
-	t.T().Helper()
-	sampleYAMLDir := getSampleYAMLDir(t)
-	sampleYAMLPath := filepath.Join(sampleYAMLDir, filename)
-	kubectlCmd := exec.CommandContext(t.Ctx(), "kubectl", "apply", "-f", sampleYAMLPath, "-n", namespace)
-	err := kubectlCmd.Run()
-	assert.NoError(t.T(), err)
-	t.T().Logf("Successfully applied %s", filename)
 }
 
 func IsPodRunningAndReady(pod *corev1.Pod) bool {
